+++ unittests/eosio_system_tester.hpp
@@ -54,9 +54,9 @@ public:
          token_abi_ser.set_abi(abi, abi_serializer_max_time);
       }
 
-      create_currency( N(eosio.token), config::system_account_name, core_from_string("10000000000.0000") );
-      issue(config::system_account_name,      core_from_string("1000000000.0000"));
-      BOOST_REQUIRE_EQUAL( core_from_string("1000000000.0000"), get_balance( "eosio" ) );
+      create_currency( N(eosio.token), config::system_account_name, core_from_string("1000000.00000000") );
+      issue(config::system_account_name,      core_from_string("100000.00000000"));
+      BOOST_REQUIRE_EQUAL( core_from_string("100000.00000000"), get_balance( "eosio" ) );
 
       set_code( config::system_account_name, contracts::eosio_system_wasm() );
       set_abi( config::system_account_name, contracts::eosio_system_abi().data() );
@@ -75,11 +75,11 @@ public:
 
       produce_blocks();
 
-      create_account_with_resources( N(alice1111111), config::system_account_name, core_from_string("1.0000"), false );
-      create_account_with_resources( N(bob111111111), config::system_account_name, core_from_string("0.4500"), false );
-      create_account_with_resources( N(carol1111111), config::system_account_name, core_from_string("1.0000"), false );
-
-      BOOST_REQUIRE_EQUAL( core_from_string("1000000000.0000"), get_balance("eosio")  + get_balance("eosio.ramfee") + get_balance("eosio.stake") + get_balance("eosio.ram") );
+      create_account_with_resources( N(alice1111111), config::system_account_name, core_from_string("0.00010000"), false );
+      create_account_with_resources( N(bob111111111), config::system_account_name, core_from_string("0.00004500"), false );
+      create_account_with_resources( N(carol1111111), config::system_account_name, core_from_string("0.00010000"), false );
+ 
+      BOOST_REQUIRE_EQUAL( core_from_string("100000.00000000"), get_balance("eosio")  + get_balance("eosio.ramfee") + get_balance("eosio.stake") + get_balance("eosio.ram") );
    }
 
    action_result open( account_name  owner,
@@ -123,8 +123,8 @@ public:
                                             mvo()
                                             ("from", creator)
                                             ("receiver", a)
-                                            ("stake_net_quantity", core_from_string("10.0000") )
-                                            ("stake_cpu_quantity", core_from_string("10.0000") )
+                                            ("stake_net_quantity", core_from_string("0.00100000") )
+                                            ("stake_cpu_quantity", core_from_string("0.00100000") )
                                             ("transfer", 0 )
                                           )
                                 );
@@ -135,7 +135,7 @@ public:
    }
 
    transaction_trace_ptr create_account_with_resources( account_name a, account_name creator, asset ramfunds, bool multisig,
-                                                        asset net = core_from_string("10.0000"), asset cpu = core_from_string("10.0000") ) {
+                                                        asset net = core_from_string("0.00100000"), asset cpu = core_from_string("0.00100000") ) {
       signed_transaction trx;
       set_transaction_headers(trx);
 
@@ -181,9 +181,9 @@ public:
       account_name creator(config::system_account_name);
       signed_transaction trx;
       set_transaction_headers(trx);
-      asset cpu = core_from_string("80.0000");
-      asset net = core_from_string("80.0000");
-      asset ram = core_from_string("1.0000");
+      asset cpu = core_from_string("0.00800000");
+      asset net = core_from_string("0.00800000");
+      asset ram = core_from_string("0.00010000");
 
       for (const auto& a: accounts) {
          authority owner_auth( get_public_key( a, "owner" ) );
@@ -437,8 +437,8 @@ public:
 
    vector<name> active_and_vote_producers() {
       //stake more than 15% of total EOS supply to activate chain
-      transfer( "eosio", "alice1111111", core_from_string("650000000.0000"), "eosio" );
-      BOOST_REQUIRE_EQUAL( success(), stake( "alice1111111", "alice1111111", core_from_string("300000000.0000"), core_from_string("300000000.0000") ) );
+      transfer( "eosio", "alice1111111", core_from_string("65000.00000000"), "eosio" );
+      BOOST_REQUIRE_EQUAL( success(), stake( "alice1111111", "alice1111111", core_from_string("30000.00000000"), core_from_string("30000.00000000") ) );
 
       // create accounts {defproducera, defproducerb, ..., defproducerz} and register as producers
       std::vector<account_name> producer_names;
@@ -470,9 +470,9 @@ public:
 
       //vote for producers
       {
-         transfer( config::system_account_name, "alice1111111", core_from_string("100000000.0000"), config::system_account_name );
-         BOOST_REQUIRE_EQUAL(success(), stake( "alice1111111", core_from_string("30000000.0000"), core_from_string("30000000.0000") ) );
-         BOOST_REQUIRE_EQUAL(success(), buyram( "alice1111111", "alice1111111", core_from_string("30000000.0000") ) );
+         transfer( config::system_account_name, "alice1111111", core_from_string("10000.00000000"), config::system_account_name );
+         BOOST_REQUIRE_EQUAL(success(), stake( "alice1111111", core_from_string("3000.00000000"), core_from_string("3000.00000000") ) );
+         BOOST_REQUIRE_EQUAL(success(), buyram( "alice1111111", "alice1111111", core_from_string("3000.00000000") ) );
          BOOST_REQUIRE_EQUAL(success(), push_action(N(alice1111111), N(voteproducer), mvo()
                                                     ("voter",  "alice1111111")
                                                     ("proxy", name(0).to_string())
@@ -501,8 +501,8 @@ public:
                                                mvo()
                                                ("from", name{config::system_account_name})
                                                ("receiver", "producer1111")
-                                               ("stake_net_quantity", core_from_string("150000000.0000") )
-                                               ("stake_cpu_quantity", core_from_string("0.0000") )
+                                               ("stake_net_quantity", core_from_string("15000.00000000") )
+                                               ("stake_cpu_quantity", core_from_string("0.00000000") )
                                                ("transfer", 1 )
                                              )
                                  );
@@ -519,8 +519,8 @@ public:
                                                mvo()
                                                ("from", "producer1111")
                                                ("receiver", "producer1111")
-                                               ("unstake_net_quantity", core_from_string("150000000.0000") )
-                                               ("unstake_cpu_quantity", core_from_string("0.0000") )
+                                               ("unstake_net_quantity", core_from_string("15000.00000000") )
+                                               ("unstake_cpu_quantity", core_from_string("0.00000000") )
                                              )
                                  );